requirements.txt
discord.py==1.7.3
Pillow==9.0.0
python_dateutil==2.8.2
requests==2.27.1
spotipy==2.19.0
